AESP-Microservice-Project/services/subscription-service/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
cryptography==41.0.0
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.0
python-dotenv==1.0.0
pika==1.3.2